trunk updates after 1.0.0.0 release
git-svn-id: http://miranda.googlecode.com/svn/trunk@3406 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/plugins/tabsrmm/chat/options.c
@@ -658,7 +658,7 @@ BOOL CALLBACK DlgProcOptions1(HWND hwndDlg,UINT uMsg,WPARAM wParam,LPARAM lParam
                             LoadMsgLogBitmaps();
                             SM_BroadcastMessage(NULL, GC_SETWNDPROPS, 0, 0, TRUE);
                         }
-                        DBWriteContactSettingByte(NULL, SRMSGMOD_T, "enable_chat", IsDlgButtonChecked(hwndDlg, IDC_CHAT_ENABLE) ? 1 : 0);
+                        DBWriteContactSettingByte(NULL, SRMSGMOD_T, "enable_chat", (BYTE)(IsDlgButtonChecked(hwndDlg, IDC_CHAT_ENABLE) ? 1 : 0));
 					}
 					return TRUE;
 				}
